new gathering and gatheringItem, functional for delete new groups (gatherings)
